PAC-73 - Extract media type loading logic into separate class
